backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.29
psycopg2-binary==2.9.9
pydantic==2.6.4
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
bcrypt==4.1.3
python-multipart==0.0.9
sendgrid==6.11.0
apscheduler==3.10.4
python-dotenv==1.0.1
httpx==0.27.0
aiofiles==23.2.1
openpyxl==3.1.2
email-validator==2.1.1
twilio==8.12.0
